+++ dependent-sum-template/src/Data/GADT/Compare/TH.hs
@@ -7,8 +7,6 @@
 module Data.GADT.Compare.TH
     ( DeriveGEQ(..)
     , DeriveGCompare(..)
-    , DeriveEqTagIdentity(..)
-    , DeriveOrdTagIdentity(..)
     , GComparing, runGComparing, geq', compare'
     ) where
 
@@ -162,99 +160,3 @@ gcompareFunction boundVars cons
                         )
                     |]
                 ) []
-
--- A type class purely for overloading purposes
-class DeriveEqTagIdentity t where
-    deriveEqTagIdentity :: t -> Q [Dec]
-
-instance DeriveEqTagIdentity Name where
-    deriveEqTagIdentity typeName = do
-        typeInfo <- reify typeName
-        case typeInfo of
-            TyConI dec -> deriveEqTagIdentity dec
-            _ -> fail "deriveEqTagIdentity: the name of a type constructor is required"
-
-instance DeriveEqTagIdentity Dec where
-    deriveEqTagIdentity = deriveForDec ''EqTag (\t -> [t| EqTag $t Identity |]) eqTaggedFunction
-
-instance DeriveEqTagIdentity t => DeriveEqTagIdentity [t] where
-    deriveEqTagIdentity [it] = deriveEqTagIdentity it
-    deriveEqTagIdentity _ = fail "deriveEqTagIdentity: [] instance only applies to single-element lists"
-
-instance DeriveEqTagIdentity t => DeriveEqTagIdentity (Q t) where
-    deriveEqTagIdentity = (>>= deriveEqTagIdentity)
-
-eqTaggedFunction bndrs cons = funD 'eqTagged $
-    map (eqTaggedClause bndrs) cons
-
-eqTaggedClause bndrs con = do
-    let argTypes = argTypesOfCon con
-        needsGEq argType = any ((`occursInType` argType) . nameOfBinder) (bndrs ++ varsBoundInCon con)
-
-    argVars <- for argTypes $ \argType ->
-        if needsGEq argType
-        then Just <$> liftA2 (,) (newName "x") (newName "y")
-        else pure Nothing
-
-    let nonWildPs = [x | Just x <- argVars]
-        doRecur = case nonWildPs of
-            [] -> Nothing
-            [var] -> Just var
-            (_:_) -> error "deriveEqTagIdentity: Can have at most one nested GADT"
-
-    clause [ conP conName (maybe wildP (varP . fst) <$> argVars)
-           , conP conName (maybe wildP (varP . snd) <$> argVars)
-           ]
-        ( normalB $ case doRecur of
-            Nothing -> [| (==) |]
-            Just (x, y) -> [| eqTagged $(varE x) $(varE y) |]
-        ) []
-    where conName = nameOfCon con
-
--- A type class purely for overloading purposes
-class DeriveOrdTagIdentity t where
-    deriveOrdTagIdentity :: t -> Q [Dec]
-
-instance DeriveOrdTagIdentity Name where
-    deriveOrdTagIdentity typeName = do
-        typeInfo <- reify typeName
-        case typeInfo of
-            TyConI dec -> deriveOrdTagIdentity dec
-            _ -> fail "deriveOrdTagIdentity: the name of a type constructor is required"
-
-instance DeriveOrdTagIdentity Dec where
-    deriveOrdTagIdentity = deriveForDec ''OrdTag (\t -> [t| OrdTag $t Identity |]) compareTaggedFunction
-
-instance DeriveOrdTagIdentity t => DeriveOrdTagIdentity [t] where
-    deriveOrdTagIdentity [it] = deriveOrdTagIdentity it
-    deriveOrdTagIdentity _ = fail "deriveOrdTagIdentity: [] instance only applies to single-element lists"
-
-instance DeriveOrdTagIdentity t => DeriveOrdTagIdentity (Q t) where
-    deriveOrdTagIdentity = (>>= deriveOrdTagIdentity)
-
-compareTaggedFunction bndrs cons = funD 'compareTagged $
-    map (compareTaggedClause bndrs) cons
-
-compareTaggedClause bndrs con = do
-    let argTypes = argTypesOfCon con
-        needsGCompare argType = any ((`occursInType` argType) . nameOfBinder) (bndrs ++ varsBoundInCon con)
-
-    argVars <- for argTypes $ \argType ->
-        if needsGCompare argType
-        then Just <$> liftA2 (,) (newName "x") (newName "y")
-        else pure Nothing
-
-    let nonWildPs = [x | Just x <- argVars]
-        doRecur = case nonWildPs of
-            [] -> Nothing
-            [var] -> Just var
-            (_:_) -> error "deriveOrdTagIdentity: Can have at most one nested GADT"
-
-    clause [ conP conName (maybe wildP (varP . fst) <$> argVars)
-           , conP conName (maybe wildP (varP . snd) <$> argVars)
-           ]
-        ( normalB $ case doRecur of
-            Nothing -> [| compare |]
-            Just (x, y) -> [| compareTagged $(varE x) $(varE y) |]
-        ) []
-    where conName = nameOfCon con